backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
tensorflow==2.15.0
numpy
opencv-python-headless
scikit-learn
